Extract resolving tests into a file
