+++ tests/Container/ContainerTest.php
@@ -335,48 +335,6 @@ class ContainerTest extends TestCase
         $this->assertEquals('taylor', $instance->default);
     }
 
-    public function testResolvingCallbacksAreCalledForSpecificAbstracts()
-    {
-        $container = new Container;
-        $container->resolving('foo', function ($object) {
-            return $object->name = 'taylor';
-        });
-        $container->bind('foo', function () {
-            return new stdClass;
-        });
-        $instance = $container->make('foo');
-
-        $this->assertEquals('taylor', $instance->name);
-    }
-
-    public function testResolvingCallbacksAreCalled()
-    {
-        $container = new Container;
-        $container->resolving(function ($object) {
-            return $object->name = 'taylor';
-        });
-        $container->bind('foo', function () {
-            return new stdClass;
-        });
-        $instance = $container->make('foo');
-
-        $this->assertEquals('taylor', $instance->name);
-    }
-
-    public function testResolvingCallbacksAreCalledForType()
-    {
-        $container = new Container;
-        $container->resolving(stdClass::class, function ($object) {
-            return $object->name = 'taylor';
-        });
-        $container->bind('foo', function () {
-            return new stdClass;
-        });
-        $instance = $container->make('foo');
-
-        $this->assertEquals('taylor', $instance->name);
-    }
-
     public function testUnsetRemoveBoundInstances()
     {
         $container = new Container;
@@ -1004,397 +962,6 @@ class ContainerTest extends TestCase
 
         $this->assertInstanceOf(ContainerImplementationStub::class, $one->impl);
         $this->assertInstanceOf(ContainerImplementationStubTwo::class, $two->impl);
-    }
-
-    public function testResolvingCallbacksShouldBeFiredWhenCalledWithAliases()
-    {
-        $container = new Container;
-        $container->alias(stdClass::class, 'std');
-        $container->resolving('std', function ($object) {
-            return $object->name = 'taylor';
-        });
-        $container->bind('foo', function () {
-            return new stdClass;
-        });
-        $instance = $container->make('foo');
-
-        $this->assertEquals('taylor', $instance->name);
-    }
-
-    public function testResolvingCallbacksAreCalledOnceForImplementation()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testGlobalResolvingCallbacksAreCalledOnceForImplementation()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testAfterResolvingCallbacksAreCalledOnceForImplementation()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->afterResolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledOnceForSingletonConcretes()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-        $container->bind(ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(3, $callCounter);
-    }
-
-    public function testResolvingCallbacksCanStillBeAddedAfterTheFirstResolution()
-    {
-        $container = new Container;
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCanceledWhenInterfaceGetsBoundToSomeOtherConcrete()
-    {
-        $container = new Container;
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $callCounter = 0;
-        $container->resolving(ContainerImplementationStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStubTwo::class);
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledOnceForStringAbstractions()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving('foo', function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind('foo', ContainerImplementationStub::class);
-
-        $container->make('foo');
-        $this->assertEquals(1, $callCounter);
-
-        $container->make('foo');
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testResolvingCallbacksForConcretesAreCalledOnceForStringAbstractions()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(ContainerImplementationStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind('foo', ContainerImplementationStub::class);
-        $container->bind('bar', ContainerImplementationStub::class);
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make('foo');
-        $this->assertEquals(2, $callCounter);
-
-        $container->make('bar');
-        $this->assertEquals(3, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(4, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledOnceForImplementation2()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, function () {
-            return new ContainerImplementationStub;
-        });
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(3, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(4, $callCounter);
-    }
-
-    public function testRebindingDoesNotAffectResolvingCallbacks()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-        $container->bind(IContainerContractStub::class, function () {
-            return new ContainerImplementationStub;
-        });
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(3, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(4, $callCounter);
-    }
-
-    public function testParametersPassedIntoResolvingCallbacks()
-    {
-        $container = new Container;
-
-        $container->resolving(IContainerContractStub::class, function ($obj, $app) use ($container) {
-            $this->assertInstanceOf(IContainerContractStub::class, $obj);
-            $this->assertInstanceOf(ContainerImplementationStubTwo::class, $obj);
-            $this->assertSame($container, $app);
-        });
-
-        $container->afterResolving(IContainerContractStub::class, function ($obj, $app) use ($container) {
-            $this->assertInstanceOf(IContainerContractStub::class, $obj);
-            $this->assertInstanceOf(ContainerImplementationStubTwo::class, $obj);
-            $this->assertSame($container, $app);
-        });
-
-        $container->afterResolving(function ($obj, $app) use ($container) {
-            $this->assertInstanceOf(IContainerContractStub::class, $obj);
-            $this->assertInstanceOf(ContainerImplementationStubTwo::class, $obj);
-            $this->assertSame($container, $app);
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStubTwo::class);
-        $container->make(IContainerContractStub::class);
-    }
-
-    public function testResolvingCallbacksAreCallWhenRebindHappenForResolvedAbstract()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStubTwo::class);
-        $this->assertEquals(2, $callCounter);
-
-        $container->make(ContainerImplementationStubTwo::class);
-        $this->assertEquals(3, $callCounter);
-
-        $container->bind(IContainerContractStub::class, function () {
-            return new ContainerImplementationStubTwo();
-        });
-        $this->assertEquals(4, $callCounter);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(5, $callCounter);
-    }
-
-    public function testRebindingDoesNotAffectMultipleResolvingCallbacks()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->resolving(ContainerImplementationStubTwo::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        // it should call the callback for interface
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        // it should call the callback for interface
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-
-        // should call the callback for the interface it implements
-        // plus the callback for ContainerImplementationStubTwo.
-        $container->make(ContainerImplementationStubTwo::class);
-        $this->assertEquals(4, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledForInterfaces()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(IContainerContractStub::class);
-
-        $this->assertEquals(1, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledForConcretesWhenAttachedOnInterface()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(ContainerImplementationStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledForConcretesWhenAttachedOnConcretes()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(ContainerImplementationStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->bind(IContainerContractStub::class, ContainerImplementationStub::class);
-
-        $container->make(IContainerContractStub::class);
-        $this->assertEquals(1, $callCounter);
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledForConcretesWithNoBinding()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(ContainerImplementationStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
-    }
-
-    public function testResolvingCallbacksAreCalledForInterFacesWithNoBinding()
-    {
-        $container = new Container;
-
-        $callCounter = 0;
-        $container->resolving(IContainerContractStub::class, function () use (&$callCounter) {
-            $callCounter++;
-        });
-
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(1, $callCounter);
-        $container->make(ContainerImplementationStub::class);
-        $this->assertEquals(2, $callCounter);
     }
 
     public function testMakeWithMethodIsAnAliasForMakeMethod()
